Add definition: educational hackathon (post-HtH)
Definition suggested during the event Hack The Hackathon (vol. 3)
related to a group work on a framework for [educational
hackathon](https://github.com/hackthackathon/educational-hackathon).

**A very basic definition to set a common ground regarding an
interpretation of the notion, could be easily challenged and improved.**

Potential definition purposes:

Hackathon can be a confusing term, educational hackathon either.

This notion and its definition may be meaningless, as the main thing is
about trying to improve practices regarding the educational dimension of
hackathon. It may be confusing by leading to the idea of hackathon with
educational purposes only.

We were not totally sure about using a specific word and we were not
focusing on a format for education only.

With this definition, the idea was to foster this interest in the
educational dimension of hackathon while having a distinct vocabulary
to discuss the topic.